modified the validations (#125)

+++ employee/src/main/java/com/pb/employee/util/PayslipUtils.java
@@ -73,13 +73,13 @@ public class PayslipUtils {
             byte[] decodedPfEmployee = Base64.getDecoder().decode(salaryRequest.getDeductions().getPfEmployee());
             byte[] decodedPfEmployer = Base64.getDecoder().decode(salaryRequest.getDeductions().getPfEmployer());
             pfc = Double.parseDouble(new String(decodedPfEmployee)) + Double.parseDouble(new String(decodedPfEmployer));
-            pfc = pfc/12.0;
+            pfc = pfc/12;
             salary.getAllowances().setPfContributionEmployee(String.valueOf(pfc));
         }
         if (salaryRequest.getAllowances().getHra() != null) {
             byte[] decodedHra = Base64.getDecoder().decode(salaryRequest.getAllowances().getHra());
             hra = Double.parseDouble(new String(decodedHra));
-            hra = ((gross/12.0)*(hra/100));
+            hra = ((gross/12)*(hra/100));
             salary.getAllowances().setHra(String.valueOf(hra));
         }
         if (salaryRequest.getAllowances().getSpecialAllowance() != null) {
@@ -118,9 +118,9 @@ public class PayslipUtils {
 
             int noOfLeaves = totalWorkingDays - noOfWorkingDays;
             if (noOfLeaves > 1) {
-                double monthlySalary = (gross / 12.0);
-                double perDaySalary = monthlySalary / totalWorkingDays;
-                lop = (noOfLeaves - 1) * perDaySalary;
+                double monthlySalary = (gross / 12);
+                int perDaySalary = (int) (monthlySalary / totalWorkingDays);
+                lop = (double) ((noOfLeaves - 1) * perDaySalary);
             }else {
                 lop = (double) 0;
             }
